+++ NEWS.md
@@ -1,3 +1,8 @@
+# elements 0.0.2 (12/04/2025)
+
+* Adding the ability to plot multiple taxa in `elements::plot_me` function.
+* Adding the argument "normalise" to the `elements::plot_me` function, which re-scales the y axis values in PDP plots using min-max scaling.
+
 # elements 0.0.1 (11/04/2025)
 
 * First working version with all modelled taxa included.